typo in libnotimpl
git-svn-id: svn://svn.openwrt.org/openwrt@9233 3c298f89-4303-0410-b956-a3cf2f4a3e73

+++ packages/libs/libnotimpl/src/math.c
@@ -119,7 +119,7 @@
 #ifdef __STDC__
     float roundf(float x)
 #else
-    float round(x)
+    float roundf(x)
 	float x;
 #endif
 {
